[CBRD-26972] Keep heap header stats private
