+++ src/storage/heap_oos.cpp
@@ -30,15 +30,11 @@
 #include "error_manager.h"
 #include "file_manager.h"
 #include "heap_file.h"
-#include "heap_file_internal.hpp"
-#include "log_append.hpp"
-#include "log_manager.h"
+#include "heap_show_scan_context.hpp"
 #include "object_representation.h"
 #include "oos_file.hpp"
 #include "oos_log.hpp"
 #include "oos_util.hpp"
-#include "recovery.h"
-#include "slotted_page.h"
 #include "storage_common.h"
 
 #include <cassert>
@@ -66,118 +62,6 @@ struct HEAP_OOS_EXPAND_STATE
   std::vector<int> vot_raw;
   std::vector<std::vector<char>> oos_blobs;
 };
-
-/*
- * heap_oos_find_vfid () - find (or optionally create) the OOS file of a heap
- *   return         : true on success, false on a genuine error (er is set)
- *   hfid (in)      : heap file identifier
- *   oos_vfid (out) : OOS file identifier; set to NULL when the heap has no OOS
- *                    file (only possible when docreate == false)
- *   docreate (in)  : if true and the OOS file does not exist, it is created and
- *                    TDE is applied to it (mirroring heap_ovf_find_vfid)
- *
- * Note: A false return ALWAYS means a real error (and er_errid () is set); it
- *   never means "no OOS file". Callers using docreate == false must inspect
- *   oos_vfid (VFID_ISNULL) to tell whether an OOS file actually exists.
- */
-bool
-heap_oos_find_vfid (THREAD_ENTRY *thread_p, const HFID *hfid, VFID *oos_vfid, bool docreate)
-{
-  HEAP_HDR_STATS *heap_hdr;	/* Header of heap structure */
-  LOG_DATA_ADDR addr_hdr;	/* Address of logging data */
-  VPID vpid;			/* Page-volume identifier */
-  RECDES hdr_recdes;		/* Header record descriptor */
-  PGBUF_LATCH_MODE mode;
-  bool success;
-
-  success = true;
-  VFID_SET_NULL (oos_vfid);
-
-  addr_hdr.vfid = &hfid->vfid;
-  addr_hdr.offset = HEAP_HEADER_AND_CHAIN_SLOTID;
-
-  /* Read the header page */
-  vpid.volid = hfid->vfid.volid;
-  vpid.pageid = hfid->hpgid;
-
-  mode = (docreate == true ? PGBUF_LATCH_WRITE : PGBUF_LATCH_READ);
-  addr_hdr.pgptr = pgbuf_fix (thread_p, &vpid, OLD_PAGE, mode, PGBUF_UNCONDITIONAL_LATCH);
-  if (addr_hdr.pgptr == NULL)
-    {
-      goto exit_on_error;
-    }
-
-#if !defined (NDEBUG)
-  (void) pgbuf_check_page_ptype (thread_p, addr_hdr.pgptr, PAGE_HEAP);
-#endif /* !NDEBUG */
-
-  /* Peek the header record */
-  if (spage_get_record (thread_p, addr_hdr.pgptr, HEAP_HEADER_AND_CHAIN_SLOTID, &hdr_recdes, PEEK) != S_SUCCESS)
-    {
-      goto exit_on_error;
-    }
-
-  heap_hdr = (HEAP_HDR_STATS *) hdr_recdes.data;
-  if (VFID_ISNULL (&heap_hdr->oos_vfid))
-    {
-      if (docreate == true)
-	{
-	  TDE_ALGORITHM tde_algo = TDE_ALGORITHM_NONE;
-
-	  /* START A TOP SYSTEM OPERATION */
-	  log_sysop_start (thread_p);
-	  if (oos_create_file (thread_p, *oos_vfid) != NO_ERROR)
-	    {
-	      log_sysop_abort (thread_p);
-	      goto exit_on_error;
-	    }
-
-	  /* Apply TDE to the new OOS file atomically with its creation.
-	   * Pattern mirrors heap_ovf_find_vfid */
-	  if (heap_get_class_tde_algorithm (thread_p, &heap_hdr->class_oid, &tde_algo) != NO_ERROR)
-	    {
-	      log_sysop_abort (thread_p);
-	      goto exit_on_error;
-	    }
-
-	  if (file_apply_tde_algorithm (thread_p, oos_vfid, tde_algo) != NO_ERROR)
-	    {
-	      log_sysop_abort (thread_p);
-	      goto exit_on_error;
-	    }
-
-	  /* Log undo, then redo */
-	  log_append_undo_data (thread_p, RVHF_STATS, &addr_hdr, sizeof (*heap_hdr), heap_hdr);
-	  VFID_COPY (&heap_hdr->oos_vfid, oos_vfid);
-	  log_append_redo_data (thread_p, RVHF_STATS, &addr_hdr, sizeof (*heap_hdr), heap_hdr);
-	  pgbuf_set_dirty (thread_p, addr_hdr.pgptr, DONT_FREE);
-
-	  log_sysop_commit (thread_p);
-	}
-      else
-	{
-	  /* No OOS file exists yet; oos_vfid stays NULL. This is not an error. */
-	  goto end;
-	}
-    }
-  else
-    {
-      VFID_COPY (oos_vfid, &heap_hdr->oos_vfid);
-    }
-
-  goto end;
-
-exit_on_error:
-  success = false;
-
-end:
-  if (addr_hdr.pgptr)
-    {
-      pgbuf_unfix_and_init (thread_p, addr_hdr.pgptr);
-    }
-
-  return success;
-}
 
 /*
  * heap_oos_parse_vot () - Walk the source VOT and collect each entry (including flag bits).
